Refactor create-teacher, book into a module
