+++ app.rb
@@ -2,10 +2,10 @@ require_relative './student'
 require_relative './teacher'
 require_relative './rental'
 require_relative './book'
-require_relative './re_student'
+require_relative './create_methods'
 
 class App
-  include ReStudent
+  include CreateMethods
 
   def initialize
     @people = []
@@ -58,60 +58,6 @@ class App
     when 6
       list_person_rentals
     end
-  end
-
-  def create_tasks(selection)
-    case selection
-    when 3
-      create_person
-    when 4
-      create_book
-    when 5
-      create_rental
-    end
-  end
-
-  def create_person
-    print 'Do you want to create a student (1) or a teacher (2)? [Input the number]: '
-    person_selected = gets.chomp.strip.to_i
-
-    case person_selected
-    when 1
-      create_student
-      @id += 1
-    when 2
-      create_teacher
-      @id += 1
-    else
-      puts 'Invalid Selection. Returning to main menu'
-      nil
-    end
-  end
-
-  def create_teacher
-    print 'Age: '
-    age = gets.chomp.to_i
-    if age <= 0
-      @id -= 1
-      return puts 'Wrong input for age. Returning to main menu'
-    end
-    print 'Name: '
-    name = gets.chomp.strip.capitalize
-    print 'Specialization: '
-    specialization = gets.chomp.strip.capitalize
-    teacher = Teacher.new(@id, age, specialization, name)
-    @people << teacher
-    puts 'Person created successfully'
-  end
-
-  def create_book
-    print 'Title: '
-    title = gets.chomp.strip.capitalize
-    print 'Author: '
-    author = gets.chomp.strip.capitalize
-    book = Book.new(title, author)
-    @books << book
-    puts 'Book created successfully'
   end
 
   def list_all_people
