+++ rental.rb
@@ -5,8 +5,6 @@ class Rental
   def initialize(date, book, person)
     @date = date
     @book = book
-    # book.rentals << self
     @person = person
-    # person.rentals << self
   end
 end
